Refactored code out of main.py into submodules
